pyproject.toml
[tool.poetry]
name = "FileZ"
version = "0.10.0"
description = "Package for abstracting file system handler in an oriented-object style."
authors = ["Gabriel <contato@gabrielfontenelle.com>"]
license = "GNU GENERAL PUBLIC LICENSE v3"
packages = [
    { include = "filez" },
]
include = [
    "filez/data/mime.types",
    "TODO",
    "LICENSE"
]

[tool.poetry.dependencies]
python = ">=3.11"
Send2Trash = "^1.5.0"
psutil = "^5.8.0"
py7zr = "^0.21"
tinytag = "1.6.0"
dill = "^0.3.5"
json-tricks = "^3.15"
rarfile = "^4.0"
tox-poetry = "^0.4"
Wand = "^0.6"
pillow = "^9.2"
opencv-python = "^4.6"
psd-tools = "^1.9"
PyPDF2 = "^2.11"
imageio = "^2.22"
pymupdf = "^1.20"
imageio-ffmpeg = "*"
av = "^11.0"
moviepy = "^1.0"
pytz = "^2022.5"
requests = "^2.31"
typing-extensions = "^4.10.0"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-shutil = "^1.7.0"
mypy = "^1.8.0"
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
